✨ Datamodel for Blocks
Minimal data that is stored on chain. This is not intended to be easy to
access data ! Only to ensure data is there and immutable. For easy &
quick access, we need to build an indexer :)

The "state cursor" is not saved on chain, you can deduce it:

If you want to know the state of a contract "A", you find the last *settled* tx that
impacted "A", and its associated "BlockProof", there you have the
current state of your contract.

How to know if a tx is settled ?
Read all next blocks starting the block where the tx is written, search
all BlockProof for this txHash of status Success, if you find such a
BlockProof for each blob, then the tx is Settled.