✨ Auto prover module (#1056)
`AutoProver` is a module that handles the proving of transactions

It listens to the node state events and processes all blobs in the
block's transactions
for a given contract.

It asynchronously generates 1 ProofTransaction to prove all concerned
blobs in a block

If a passed BlobTransaction times out, or is settled as failed, all
blobs that are "after" the failed transaction in the block are
re-executed, and prooved all at once, even if in multiple blocks.

This module requires the ELF to support multiproof. i.e. it requires the
ELF to read a `Vec<Calldata>` as input.